requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic>=2.5.2
python-dotenv
pymysql
passlib[bcrypt]
python-jose
fpdf
python-multipart
email-validator
flask
psycopg2-binary
bcrypt==4.0.1
easyocr
torch
torchvision
Pillow